Add long press detection to protocol row [RCORE-599]
