+++ app/src/pages/OnDeviceDisplay/ProtocolDashboard/index.tsx
@@ -1,9 +1,7 @@
 import * as React from 'react'
 import { useDispatch, useSelector } from 'react-redux'
-import { useHistory } from 'react-router-dom'
 import { useTranslation } from 'react-i18next'
 import styled from 'styled-components'
-import { format, formatDistance } from 'date-fns'
 import {
   COLORS,
   SPACING,
@@ -11,7 +9,6 @@ import {
   Flex,
   DIRECTION_COLUMN,
   JUSTIFY_SPACE_BETWEEN,
-  truncateString,
   Btn,
   Icon,
   JUSTIFY_CENTER,
@@ -28,6 +25,7 @@ import {
   getProtocolsOnDeviceSortKey,
   updateConfigValue,
 } from '../../../redux/config'
+import { ProtocolRow } from './ProtocolRow'
 import { sortProtocols } from './utils'
 
 import imgSrc from '../../../assets/images/odd/abstract@x2.png'
@@ -51,22 +49,10 @@ const TableHeader = styled('th')`
   font-size: ${TYPOGRAPHY.fontSizeCaption};
   padding: ${SPACING.spacing2};
 `
-const TableRow = styled('tr')`
-  border: 1px ${COLORS.medGreyHover} solid;
-  height: 4rem;
-  padding: ${SPACING.spacing5};
-`
-
-const TableDatum = styled('td')`
-  padding: ${SPACING.spacing2};
-  white-space: break-spaces;
-  text-overflow: wrap;
-`
 
 export function ProtocolDashboard(): JSX.Element {
   const protocols = useAllProtocolsQuery()
   const runs = useAllRunsQuery()
-  const history = useHistory()
   const { t } = useTranslation('protocol_info')
   const dispatch = useDispatch<Dispatch>()
   const sortBy = useSelector(getProtocolsOnDeviceSortKey) ?? 'alphabetical'
@@ -192,46 +178,13 @@ export function ProtocolDashboard(): JSX.Element {
                 run => run.protocolId === protocol.id
               )?.createdAt
 
-              const protocolName =
-                protocol.metadata.protocolName ?? protocol.files[0].name
-
               return (
-                <TableRow
-                  key={protocol.key ?? index}
-                  onClick={() => history.push(`/protocols/${protocol.id}`)}
-                >
-                  <TableDatum>
-                    <StyledText
-                      fontSize="1.5rem"
-                      lineHeight="2.0625rem"
-                      fontWeight={TYPOGRAPHY.fontWeightSemiBold}
-                    >
-                      {truncateString(protocolName, 88, 66)}
-                    </StyledText>
-                  </TableDatum>
-                  <TableDatum>
-                    <StyledText
-                      fontSize="1.375rem"
-                      lineHeight="1.75rem"
-                      fontWeight={TYPOGRAPHY.fontWeightRegular}
-                    >
-                      {lastRun != null
-                        ? formatDistance(new Date(lastRun), new Date(), {
-                            addSuffix: true,
-                          })
-                        : t('no_history')}
-                    </StyledText>
-                  </TableDatum>
-                  <TableDatum>
-                    <StyledText
-                      fontSize="1.375rem"
-                      lineHeight="1.75rem"
-                      fontWeight={TYPOGRAPHY.fontWeightRegular}
-                    >
-                      {format(new Date(protocol.createdAt), 'Pp')}
-                    </StyledText>
-                  </TableDatum>
-                </TableRow>
+                <ProtocolRow
+                  index={index}
+                  key={index}
+                  lastRun={lastRun}
+                  protocol={protocol}
+                />
               )
             })}
           </tbody>
@@ -244,7 +197,7 @@ export function ProtocolDashboard(): JSX.Element {
           height="27.75rem"
           backgroundColor={COLORS.medGreyEnabled}
         >
-          <img src={imgSrc} />
+          <img title={t('nothing_here_yet')} src={imgSrc} />
           <StyledText
             fontSize="2rem"
             lineHeight="2.75rem"
